+++ src/Core/Restaurant.Domain/Entities/Catalog/Ingredient.cs
@@ -10,12 +10,13 @@ namespace Restaurant.Domain.Entities.Catalog
         public string Name { get; private set; } = string.Empty;
         public string? Description { get; private set; }
 
-        public int BaseUnitId { get; private set; }
-        public int CategoryId { get; private set; }
-        public int? BrandId { get; private set; }
+        public long BaseUnitId { get; private set; }
+        public long CategoryId { get; private set; }
+        public long? BrandId { get; private set; }
 
         public Unit BaseUnit { get; private set; } = null!;
         public Brand? Brand { get; private set; }
+
         public IngredientCategory IngredientCategory { get; private set; } = null!;
         public IngredientPrice IngredientPrice { get; private set; } = null!;
         public ICollection<IngredientStock> IngredientStocks { get; private set; } = [];
@@ -25,19 +26,19 @@ namespace Restaurant.Domain.Entities.Catalog
 
     public partial class Ingredient
     {
-        public Ingredient SetBrand(int? brandId)
+        public Ingredient SetBrand(long? brandId)
         {
             BrandId = brandId;
             return this;
         }
 
-        public Ingredient SetCategory(int categoryId)
+        public Ingredient SetCategory(long categoryId)
         {
             CategoryId = categoryId;
             return this;
         }
 
-        public Ingredient SetUnit(int unitId)
+        public Ingredient SetUnit(long unitId)
         {
             BaseUnitId = unitId;
             return this;
