+++ src/Core/Restaurant.Domain/Entities/Guest/WalletTransaction.cs
@@ -5,7 +5,7 @@ namespace Restaurant.Domain.Entities.Guest
 {
     public class WalletTransaction : SoftDeletableEntity
     {
-        public int WalletId { get; private set; }
+        public long WalletId { get; private set; }
 
         public decimal Amount { get; private set; }
 
@@ -19,7 +19,7 @@ namespace Restaurant.Domain.Entities.Guest
 
         public string? Description { get; private set; }
 
-        public int? ReferenceId { get; private set; }
+        public long? ReferenceId { get; private set; }
         public WalletReferenceType? ReferenceType { get; private set; }
 
         public Wallet Wallet { get; private set; } = null!;
