+++ src/Core/Restaurant.Domain/Entities/Pricing/IngredientPrice.cs
@@ -8,16 +8,14 @@ namespace Restaurant.Domain.Entities.Pricing
         public decimal UnitPrice { get; private set; }
         public string Currency { get; private set; } = null!;
 
-        public int IngredientId { get; private set; }
-
-
+        public long IngredientId { get; private set; }
 
         public Ingredient Ingredient { get; private set; } = null!;
     }
 
     public partial class IngredientPrice
     {
-        public IngredientPrice SetIngredient(int ingredientId)
+        public IngredientPrice SetIngredient(long ingredientId)
         {
             IngredientId = ingredientId;
             return this;
